feat: titleArea, textArea에 selectRecord 값 가져오기
