+++ src/components/titleArea/index.jsx
@@ -1,8 +1,20 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
+import useSelectRecordStore from "../../store/selectRecord";
 import TitleArea from "./TitleArea";
 
 const TitleAreaComponent = () => {
-  return <TitleArea></TitleArea>;
+  const selectRecord = useSelectRecordStore((state) => state.selectRecord);
+  const [record, setRecord] = useState(selectRecord);
+
+  useEffect(() => {
+    updateRecord(selectRecord);
+  }, [selectRecord]);
+
+  const updateRecord = (r) => {
+    setRecord(r);
+  };
+
+  return <TitleArea record={record}></TitleArea>;
 };
 
 export default TitleAreaComponent;
